+++ demo/config/view.js
@@ -37,7 +37,6 @@ module.exports = {
   /**
    * Hook function just before the view is rendered.
    * For example, you can set your own local variables that can be used within the view.
-   *
    * @example
    * // The message set here can be referenced in the view as {{message}}.
    * beforeRender: (req, res) => {