+++ demo/routes/users.js
@@ -1,7 +1,7 @@
 const express = require('express');
-const router = express.Router();
 const UserModel = require('../models/UserModel');
 
+const router = express.Router();
 router.get('/', (req, res, next) => {
   res.render('users');
 });