+++ src/utils/loadDatabaseConfig.ts
@@ -1,14 +1,13 @@
 import fs from 'node:fs';
-import {createRequire} from 'node:module';
 import sequelize from 'sequelize';
 import DatabaseConfig from '~/interfaces/DatabaseConfig';
-const require = createRequire(import.meta.url);
+import Environment from '~/middlewares/Environment';
 
 /**
   * Get database configuration (config/database).
-  * @return {sequelize.Options} Loaded configuration.
+  * @return {Promise<sequelize.Options>} Loaded configuration.
   */
-export default (): sequelize.Options => {
+export default async (): Promise<sequelize.Options> => {
   // Options with default values set.
   const defaultOptions: sequelize.Options = {
     database: 'unkown',
@@ -23,13 +22,13 @@ export default (): sequelize.Options => {
   if (!fs.existsSync(`${filePath}.js`))
     return defaultOptions;
 
-  // Get the execution environment.
-  const env = process.env.NODE_ENV||'development';
+  // Get NODE_ENV environment variable.
+  await Environment.mount();
 
   // If an options file is found, it returns options that override the default options.
-  const options = <DatabaseConfig>require(filePath).default||require(filePath);
-  const mergeOptions = Object.assign(defaultOptions, options[env]);
+  let {default: options}: {default: DatabaseConfig} = await import(`${filePath}.js`);
+  options = Object.assign(defaultOptions, options[process.env.NODE_ENV||'development']) as DatabaseConfig;
   if (process.env.EXPRESS_DEBUG)
-    console.log(`Connection DB is ${mergeOptions.host} ${mergeOptions.database}`);
-  return mergeOptions;
+    console.log(`Connection DB is ${options.host} ${options.database}`);
+  return options;
 }